backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pgvector
sentence-transformers
python-dotenv
pypdf
